requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Audio/Video processing
whisper-openai==20231117
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0

# AI and NLP
openai==1.3.7
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Vector Database and Graph Processing
weaviate-client==3.25.3
graphrag==0.1.5
neo4j==5.14.1
networkx==3.2.1

# Enhanced NLP and Web Processing
nlweb==0.3.2
spacy==3.7.2
beautifulsoup4==4.12.2
scrapy==2.11.0

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0

# PDF and document generation
reportlab==4.0.7
markdown==3.5.1
jinja2==3.1.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2

# Web Research and Knowledge Graph
networkx==3.2.1
spacy==3.7.2
weaviate-client==3.25.3
nlweb==0.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
